fix(directory):
The issue was twofold, and related to the assumptions made around
directories in the plugin. As it turns out, when creating a directory in
Nextcloud 28, Sabre contacts our plugin in two distinct places:
getMetaData and file_exists. While the getMetaData call succeeds and
populates the MKCOL response, the existence of an oc-fileid HTTP header
depends on file_exists, which will always return false for directories.

This fix involves adding directories as files without Swarm-specific
metadata to the Swarm file mapping database. This approach may introduce
incompatibilities, and require a Migration step to be devised to
populate the table and get rid of the cache-based assumption.

This commit does not remove the cache-based assumption so as to not
break existing file databases.