fix(copyref): Copy Swarm Reference issue
This is a fix for the Copy Swarm Reference feature. Due to the requirement
for a NPM module @nextcloud/files, the JS feature has been moved to src/
and added to the Webpack build list.

As such, `npm run build` is required to build the relevant JS files and
preview this fix.