feature: file actions + archive + installation + bugfixes + dev env (#126)
I tested the following features in the local development environment (NC 30):
- Comments: Working properly.
- Tags: Working properly.
- Upload: Working properly.
- Download: Working properly.
- Archive Files and Folders: Working properly.
- Unarchive Files and Folders: Working properly.
- Rename: Working properly.

However, there is one area of concern regarding the activation of the plug-in when external storage is not enabled. I encountered some issues in the local development environment, but I will assume (and hope) that these problems are specific to this environment and will not occur in production.